+++ src/__tests__/App.scale.test.tsx
@@ -41,6 +41,7 @@ vi.mock('../chart/adapter', async (importOriginal) => {
     setReferencePrice() {}
     setMarkerPrice() {}
     setSessionHighLow() {}
+    setTradeMarkers() {}
     setPositionDragHandler() {}
     setDrawings() {}
     setDrawingTool() {}